requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
uwsgi
requests
pint
PyCrypto
cassandra-driver
pandas